:memo: Make Verification Work
It doesn't appear that people run the design verification stuff
much, or at all?  It has some basic functionality now but I'm not
sure if it is working 'correctly' since the resulting logfiles do
not match the archived logfiles.  So this functionality could be
improved to be sure that it is working correctly, and that
coverage is complete.