:memo: Update comments to avoid defines
Closes #3
This moves comments from the end of defines to be on the line
prior to the define.  The initial request for this was to use
the comment style with a close /* this style */  but it seems
that style wasn't available in the intial verilog language.

A new enhancement will be created to update all code to a later
verilog version.